+++ bindings/nostr-sdk-ffi/src/protocol/nips/nip34.rs
@@ -12,7 +12,6 @@ use nostr::{RelayUrl, Url};
 use uniffi::{Enum, Record};
 
 use crate::error::NostrSdkError;
-use crate::protocol::event::EventId;
 use crate::protocol::key::PublicKey;
 use crate::protocol::nips::nip01::Coordinate;
 use crate::protocol::types::Timestamp;
@@ -186,16 +185,14 @@ impl TryFrom<GitPatchContent> for nip34::GitPatchContent {
 /// Git Patch
 #[derive(Record)]
 pub struct GitPatch {
-    /// Repository ID
-    pub repo_id: String,
+    /// Repository
+    pub repository: Arc<Coordinate>,
     /// Patch
     pub content: GitPatchContent,
-    /// Maintainers
-    pub maintainers: Vec<Arc<PublicKey>>,
     /// Earliest unique commit ID of repo
     pub euc: String,
-    /// Root proposal ID
-    pub root_proposal_id: Option<Arc<EventId>>,
+    /// Labels
+    pub labels: Vec<String>,
 }
 
 impl TryFrom<GitPatch> for nip34::GitPatch {
@@ -203,11 +200,10 @@ impl TryFrom<GitPatch> for nip34::GitPatch {
 
     fn try_from(value: GitPatch) -> Result<Self, Self::Error> {
         Ok(Self {
-            repo_id: value.repo_id,
+            repository: value.repository.as_ref().deref().clone(),
             content: value.content.try_into()?,
-            maintainers: value.maintainers.into_iter().map(|p| **p).collect(),
-            euc: value.euc,
-            root_proposal_id: value.root_proposal_id.map(|e| **e),
+            euc: Sha1Hash::from_str(&value.euc)?,
+            labels: value.labels,
         })
     }
 }